% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-wp.R
\name{read.wp}
\alias{read.wp}
\title{read.wp}
\usage{
read.wp(file)
}
\arguments{
\item{file}{wikipathway gmt file
downloaded from 'https://wikipathways-data.wmcloud.org/current/gmt/'}
}
\value{
a 'gson' object
}
\description{
parse wikipathway gmt file to a gson object
}
\author{
Guangchuang Yu
}
