% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpplot.R
\name{wp_shadowtext}
\alias{wp_shadowtext}
\title{Add halo above gene name to get a clear view.}
\usage{
wp_shadowtext(p, bg.r = 2, bg.col = "white")
}
\arguments{
\item{p}{An wpplot class variance.}

\item{bg.r}{The width of halo.}

\item{bg.col}{The color of halo.}
}
\value{
A 'wpplot' object
}
\description{
Add use svghalo2 function to add halo.
}
