% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpplot.R
\name{wpplot}
\alias{wpplot}
\title{Input specific wikipathways ID to get an output in class of wpplot.}
\usage{
wpplot(ID)
}
\arguments{
\item{ID}{ID is wikipathways' ID.}
}
\value{
A 'wpplot' object
}
\description{
Use wikipathways ID to open a local svg file. Then extract related information from svg file and build a wpplot class variance.
}
\examples{
\dontrun{
   wpplot('WP179') 
}
}
