% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpplot.R
\name{wpsave}
\alias{wpsave}
\title{Save svg picture in specific address.}
\usage{
wpsave(p, file, width = NULL, height = NULL)
}
\arguments{
\item{p}{A wpplot class variance.}

\item{file}{Address where you want to save the svg.}

\item{width}{Width of svg.}

\item{height}{Height of svg.}
}
\description{
Save svg picture in specific address.
}
