% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiprofiler-package.R
\docType{package}
\name{wikiprofiler-package}
\alias{wikiprofiler}
\alias{wikiprofiler-package}
\title{wikiprofiler: Data Integration and Visualization on 'WikiPathways' Graphics}
\description{
Retrieves pathway graphics from 'WikiPathways' and maps user-supplied quantitative data, such as gene expression values, onto pathway nodes. Provides a pipe-friendly grammar for building pathway visualizations with layered fills, text highlighting, condition-wise comparisons, data preparation helpers, and batch rendering utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/contribution-knowledge-mining/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/wikiprofiler/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]
  \item Yihao Chen \email{olivercyh@foxmail.com}
}

}
\keyword{internal}
