% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiprofiler-package.R
\docType{package}
\name{wikiprofiler-package}
\alias{wikiprofiler}
\alias{wikiprofiler-package}
\title{wikiprofiler: 'WikiPathway' Based Data Integration and Visualization}
\description{
Queries online 'WikiPathway' graphics and allows mapping user data (e.g., expression values) on the graph. The package designs a grammar of graphic syntax that using pipe operator to add graphic layer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/contribution-knowledge-mining/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/wikiprofiler/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Yihao Chen \email{olivercyh@foxmail.com}
}

}
\keyword{internal}
