% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpplot.R
\name{wpsave}
\alias{wpsave}
\title{Save the 'wpplot' object to a file.}
\usage{
wpsave(p, file, width = NULL, height = NULL, ...)
}
\arguments{
\item{p}{A 'wpplot' object}

\item{file}{the file to save the object}

\item{width}{Width of the figure}

\item{height}{Height of the figure}

\item{...}{additional parameter passed to 'ggsave'}
}
\value{
output the file and the input 'wpplot' object (invisible)
}
\description{
Save the 'wpplot' object to a file.
}
